# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ShelfOps API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/incidents": {
      "get": {
        "summary": "List visible incidents",
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "query",
            "name": "cursor",
            "required": false,
            "description": "Opaque continuation cursor; reuse only with the same session subject, filters, and fixed updated-time order."
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "query",
            "name": "storeId",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "query",
            "name": "sectorId",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "query",
            "name": "locationId",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "open",
                "classified",
                "in-progress",
                "blocked",
                "resolved"
              ]
            },
            "in": "query",
            "name": "state",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "query",
            "name": "category",
            "required": false,
            "description": "Exact catalog category key; unknown or nonmatching keys return an empty authorized result."
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "query",
            "name": "severity",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "query",
            "name": "assigneeId",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "query",
            "name": "reporterId",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "on-track",
                "warning",
                "breached",
                "paused"
              ]
            },
            "in": "query",
            "name": "slaCondition",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "confirmed",
                "dismissed"
              ]
            },
            "in": "query",
            "name": "recurrenceDecisionState",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "in": "query",
            "name": "createdFrom",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "in": "query",
            "name": "createdTo",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "in": "query",
            "name": "updatedFrom",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "in": "query",
            "name": "updatedTo",
            "required": false
          }
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "headers": {
              "X-Correlation-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Server-generated identifier for support correlation."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Incident",
                        "unevaluatedProperties": false
                      }
                    },
                    "limit": {
                      "type": "integer"
                    },
                    "nextCursor": {
                      "type": "string",
                      "description": "Opaque continuation cursor present only when more rows exist."
                    },
                    "correlationId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "items",
                    "limit",
                    "correlationId"
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/paths/~1api~1v1/get/responses/400"
          },
          "401": {
            "$ref": "#/paths/~1api~1v1~1stores~1{storeId}~1locations~1{locationId}~1configuration/post/responses/401"
          },
          "503": {
            "$ref": "#/paths/~1api~1v1/get/responses/503"
          }
        }
      },
      "post": {
        "summary": "Create an incident with initial text evidence",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "storeId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "sectorId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "locationId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "productId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "category": {
                    "type": "string",
                    "pattern": "\\S"
                  },
                  "severity": {
                    "type": "string",
                    "pattern": "\\S"
                  },
                  "title": {
                    "type": "string",
                    "pattern": "\\S"
                  },
                  "description": {
                    "type": "string",
                    "pattern": "\\S"
                  },
                  "occurredAt": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "textEvidence": {
                    "type": "string",
                    "pattern": "\\S",
                    "maxLength": 4000
                  },
                  "idempotencyKey": {
                    "type": "string",
                    "pattern": "\\S"
                  }
                },
                "required": [
                  "storeId",
                  "sectorId",
                  "locationId",
                  "category",
                  "severity",
                  "title",
                  "description",
                  "occurredAt",
                  "textEvidence",
                  "idempotencyKey"
                ]
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "headers": {
              "X-Correlation-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Correlation for this HTTP request; replays receive a new value."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "incidentId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "evidenceId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "eventId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "reporterId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "state": {
                      "type": "string",
                      "enum": [
                        "open"
                      ]
                    },
                    "version": {
                      "type": "integer",
                      "enum": [
                        1
                      ]
                    },
                    "correlationId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "incidentId",
                    "evidenceId",
                    "eventId",
                    "reporterId",
                    "createdAt",
                    "state",
                    "version",
                    "correlationId"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "headers": {
              "X-Correlation-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Correlation for this HTTP request; replays receive a new value."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "code",
                    "message",
                    "correlationId",
                    "fields"
                  ],
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "validation-failed"
                      ]
                    },
                    "message": {
                      "type": "string"
                    },
                    "correlationId": {
                      "description": "Server-generated identifier for support correlation.",
                      "type": "string"
                    },
                    "fields": {
                      "type": "array",
                      "items": {
                        "additionalProperties": false,
                        "type": "object",
                        "required": [
                          "name",
                          "code"
                        ],
                        "properties": {
                          "name": {
                            "type": "string"
                          },
                          "code": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "headers": {
              "X-Correlation-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Correlation for this HTTP request; replays receive a new value."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "code",
                    "message",
                    "correlationId"
                  ],
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "authentication-required"
                      ]
                    },
                    "message": {
                      "type": "string"
                    },
                    "correlationId": {
                      "description": "Server-generated identifier for support correlation.",
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "headers": {
              "X-Correlation-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Correlation for this HTTP request; replays receive a new value."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "code",
                    "message",
                    "correlationId"
                  ],
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "forbidden"
                      ]
                    },
                    "message": {
                      "type": "string"
                    },
                    "correlationId": {
                      "description": "Server-generated identifier for support correlation.",
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "headers": {
              "X-Correlation-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Correlation for this HTTP request; replays receive a new value."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "code",
                    "message",
                    "correlationId"
                  ],
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "idempotency-conflict"
                      ]
                    },
                    "message": {
                      "type": "string"
                    },
                    "correlationId": {
                      "description": "Server-generated identifier for support correlation.",
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "headers": {
              "X-Correlation-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Correlation for this HTTP request; replays receive a new value."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "code",
                    "message",
                    "correlationId"
                  ],
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "temporarily-unavailable"
                      ]
                    },
                    "message": {
                      "type": "string"
                    },
                    "correlationId": {
                      "description": "Server-generated identifier for support correlation.",
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "shelfops_session"
      }
    },
    "schemas": {
      "Incident": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "storeId": {
            "type": "string",
            "format": "uuid"
          },
          "sectorId": {
            "type": "string",
            "format": "uuid"
          },
          "locationId": {
            "type": "string",
            "format": "uuid"
          },
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "category": {
            "type": "string",
            "description": "Category key from the organization catalog."
          },
          "severity": {
            "type": "string"
          },
          "categoryProvisional": {
            "type": "boolean"
          },
          "severityProvisional": {
            "type": "boolean"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "occurredAt": {
            "type": "string",
            "format": "date-time"
          },
          "reporterId": {
            "type": "string",
            "format": "uuid"
          },
          "assigneeId": {
            "type": "string",
            "format": "uuid"
          },
          "ownershipGap": {
            "type": "boolean"
          },
          "state": {
            "type": "string",
            "enum": [
              "open",
              "classified",
              "in-progress",
              "blocked",
              "resolved"
            ]
          },
          "reopenCount": {
            "type": "integer"
          },
          "version": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "storeId",
          "sectorId",
          "locationId",
          "category",
          "severity",
          "categoryProvisional",
          "severityProvisional",
          "title",
          "description",
          "occurredAt",
          "reporterId",
          "ownershipGap",
          "state",
          "reopenCount",
          "version",
          "createdAt",
          "updatedAt"
        ]
      }
    }
  }
}